requirements.txt
cartopy>=0.22
matplotlib>=3.8
numpy>=1.24
rasterio>=1.3
geopandas>=0.14
shapely>=2.0
requests>=2.31
Pillow>=10.0
pyproj>=3.6
fiona>=1.9
adjustText>=1.1
tqdm>=4.66
genanki>=0.13
pytest>=7.0
